+++ src/lib.rs
@@ -44,21 +44,25 @@ async-trait = "0.1"
 ```
 
 ```no_run
+use async_trait::async_trait;
+use aws_sdk_s3::operation::get_object::GetObjectOutput;
+use std::collections::HashMap;
+
 use s3sync::config::Config;
 use s3sync::config::args::parse_from_args;
 use s3sync::pipeline::Pipeline;
 use s3sync::types::SyncStatistics;
 use s3sync::types::event_callback::{EventCallback, EventData, EventType};
+use s3sync::types::preprocess_callback::{PreprocessCallback, PreprocessError, UploadMetadata};
 use s3sync::types::token::create_pipeline_cancellation_token;
 
-use async_trait::async_trait;
-
+// This struct represents a user-defined event callback.
+// You can use this callback to handle events, such as logging, monitoring, or custom actions.
 pub struct DebugEventCallback;
 
 #[async_trait]
 impl EventCallback for DebugEventCallback {
     // A callback function is called when an event occurs in the pipeline.
-    // You can use this callback to handle events, such as logging, monitoring, or custom actions.
     // The callbacks are called serially, and the callback function MUST return immediately.
     // If a callback function takes a long time to execute, it may block a whole pipeline.
     async fn on_event(&mut self, event_data: EventData) {
@@ -69,10 +73,46 @@ impl EventCallback for DebugEventCallback {
             EventType::SYNC_COMPLETE => {
                 println!("Sync complete: {event_data:?}");
             }
+            EventType::SYNC_CANCEL => {
+                println!("Sync cancelled: {event_data:?}");
+            }
             _ => {
                 println!("Other events: {event_data:?}");
             }
         }
+    }
+}
+
+// This struct represents a user-defined preprocessed callback.
+// It can be used to implement custom preprocessing logic before uploading objects to S3.
+pub struct DebugPreprocessCallback;
+
+#[async_trait]
+impl PreprocessCallback for DebugPreprocessCallback {
+    // The callbacks are called serially, and the callback function MUST return immediately.
+    // If a callback function takes a long time to execute, it may block a whole pipeline.
+    async fn preprocess_before_upload(
+        &mut self,
+        key: &str,                       // The key of the object being uploaded
+        source_object: &GetObjectOutput, // The source object being uploaded(read only)
+        metadata: &mut UploadMetadata,   // The metadata for the upload, which can be modified
+    ) -> anyhow::Result<()> {
+        // If we want to cancel the upload, return an error with PreprocessError::Cancelled
+        if key == "callback_cancel_test" || key == "data1" {
+            return Err(anyhow::Error::from(PreprocessError::Cancelled));
+        }
+
+        // The following code is an example of how to modify the metadata before uploading based on the source object.
+        let content_length = source_object.content_length.unwrap().to_string();
+        if let Some(user_defined_metadata) = metadata.metadata.as_mut() {
+            user_defined_metadata.insert("mycontent-length".to_string(), content_length);
+        } else {
+            let mut user_defined_metadata = HashMap::new();
+            user_defined_metadata.insert("mycontent-length".to_string(), content_length);
+            metadata.metadata = Some(user_defined_metadata);
+        }
+
+        Ok(())
     }
 }
 
@@ -94,9 +134,15 @@ async fn main() {
     // You can specify the events you want to capture. E.g. `EventType::SYNC_START | EventType::SYNC_COMPLETE`
     // If you want to capture all events, use `EventType::ALL_EVENTS`.
     config.event_manager.register_callback(
-        EventType::SYNC_START | EventType::SYNC_COMPLETE,
+        EventType::SYNC_START | EventType::SYNC_COMPLETE | EventType::SYNC_CANCEL,
         DebugEventCallback {},
     );
+
+    // This is a preprocess manager that manages the preprocess callbacks.(optional)
+    // You can register a preprocess callback to dynamically modify the upload metadata before uploading objects to S3.
+    config
+        .preprocess_manager
+        .register_callback(DebugPreprocessCallback {});
 
     // Create a cancellation token for the pipeline.
     // You can use this token to cancel the pipeline.